backups/scripts/requirements.txt
# Backup Scripts Dependencies
# ===========================
# Install with: pip install -r requirements.txt

# AWS S3-compatible API (for Cloudflare R2)
boto3>=1.34.0

# PostgreSQL client library (for verification)
psycopg2-binary>=2.9.9

# Environment variables
python-dotenv>=1.0.0

# HTTP client (for Telegram notifications)
httpx>=0.27.0

# Note: pg_dump and pg_restore must be installed separately
# - Windows: Install PostgreSQL or pgAdmin
# - Mac: brew install postgresql
# - Linux: apt install postgresql-client
